Added fastqc module
